+++ src/ModularPipelines.Build/Modules/CodeFormattedNicelyModule.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Options;
+using ModularPipelines.Attributes;
 using ModularPipelines.Build.Settings;
 using ModularPipelines.Context;
 using ModularPipelines.DotNet.Extensions;
@@ -9,6 +10,7 @@ using ModularPipelines.Modules;
 
 namespace ModularPipelines.Build.Modules;
 
+[SkipIfDependabot]
 public class CodeFormattedNicelyModule : Module<CommandResult>
 {
     private readonly IOptions<GitHubSettings> _githubSettings;