+++ src/ModularPipelines/Attributes/MandatoryRunConditionAttribute.cs
@@ -1,0 +1,5 @@
+namespace ModularPipelines.Attributes;
+
+public abstract class MandatoryRunConditionAttribute : RunConditionAttribute
+{
+}