+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,3 +1,4 @@
 *   Equality compareres for File and Folder objects based on their paths
 *   Added pipeline runners for Windows and Mac, alongside the existing Linux
 *   Module Attributes RunOnLinux, RunOnMacOS, RunOnWindows - To run certain modules only on certain operating systems
+*   Skip if Dependabot attribute